+++ CompanionAnimalProject/src/main/java/com/example/demo/controller/BoardController.java
@@ -174,34 +174,34 @@ public class BoardController {
 	public String myPage() {
 		return "board/community/myPage";
 	}
-	
-	/* 내가 쓴 글 확인 */
-	@GetMapping(value = "/community/myPosts")
-    public String myPosts(HttpSession session, Model model) throws Exception {
-		
-		User userInfo = (User) session.getAttribute("userInfo");
-		String id = userInfo.getId();
+//	
+//	/* 내가 쓴 글 확인 */
+//	@GetMapping(value = "/mypage/myPosts")
+//    public String myPosts(HttpSession session, Model model) throws Exception {
+//		
+//		User userInfo = (User) session.getAttribute("userInfo");
+//		String id = userInfo.getId();
+//
+//		List<Board> userPosts = boardService.selectByUserId(id);
+//		
+//        model.addAttribute("boardList", userPosts);
+//    
+//        return "mypage/myPosts";
+//    }
 
-		List<Board> userPosts = boardService.selectByUserId(id);
-		
-        model.addAttribute("boardList", userPosts);
-    
-        return "board/community/myPosts";
-    }
-
 	
 	
-	// 동물병원 메인 페이지
-	@GetMapping(value = "/hospital/main")
-	public static String hospitalMain(Model model) {
-		model.addAttribute("test", "게시글 작성 페이지");
-        return "board/hospital/main";
-	}
-
-	// 동물 서비스 메인 페이지
-	@GetMapping(value = "/service/main")
-	public static String serviceMain() {
-        return "board/service/main";
-	}
+//	// 동물병원 메인 페이지
+//	@GetMapping(value = "/hospital/main")
+//	public static String hospitalMain(Model model) {
+//		model.addAttribute("test", "게시글 작성 페이지");
+//        return "board/hospital/main";
+//	}
+//
+//	// 동물 서비스 메인 페이지
+//	@GetMapping(value = "/service/main")
+//	public static String serviceMain() {
+//        return "board/service/main";
+//	}
 	
 }
