+++ CompanionAnimalProject/src/main/webapp/WEB-INF/views/fragments/nav.jsp
@@ -39,14 +39,16 @@
                         <a class="nav-link" href="<c:url value="/dictionary/main"/>">동물보감</a>
                     </li>
 
-					<c:if test="${userInfo ==null }">
+					<c:if test="${userInfo == null }">
                     <li class="nav-item ms-3">
-                        <a class="nav-link custom-btn custom-border-btn btn" href="<c:url value="/member/login"/>">로그인</a>
+                        <a class="nav-link custom-btn custom-border-btn btn" href="<c:url value="/member/login"/>">LOGIN</a>
                     </li>
                     </c:if>
-                    <c:if test="${userInfo !=null }">
-                    <a class="nav-link custom-btn custom-border-btn btn" href="<c:url value="/board/community/myPage"/>">마이페이지</a>
-                    <a class="nav-link custom-btn custom-border-btn btn" href="<c:url value="/member/logout"/>">로그아웃</a>                        	
+                    <c:if test="${userInfo != null }">
+                 	<li class="nav-item dropdown">
+                        <a class="nav-link" href="<c:url value="/mypage/main"/>"><b>MY PAGE</b></a>
+                    </li>
+                   	<a class="nav-link custom-btn custom-border-btn btn" href="<c:url value="/member/logout"/>">LOGOUT</a>                        	
 					</c:if>
                 </ul>
             </div>
