+++ CompanionAnimalProject/src/main/webapp/WEB-INF/views/main.jsp
@@ -20,16 +20,7 @@
 	<%@include file="fragments/common-css.jsp" %>
 	
 </head>
-<body id="section_1"> <br>
-아이디 : ${userInfo.id }<br>
-<c:if test="${userInfo ==null }">
-	<input type="button" value="로그인" onclick="location.href='./member/login'">
-</c:if>
-<c:if test="${userInfo !=null }">
-	<a href="<c:url value="/member/logout"/>"> 로그아웃</a>
-	<a href="<c:url value="/member/modify"/>"> 비밀번호 변경</a>
-	<a href="<c:url value="/board/community/myPage"/>"> 마이페이지</a>                           	
-</c:if>
+<body id="section_1">
 	<!-- header -->
 	<%@include file="fragments/header.jsp" %>
