+++ CompanionAnimalProject/src/main/webapp/WEB-INF/views/member/modify.jsp
@@ -6,8 +6,20 @@
 <!DOCTYPE html>
 <html>
 <head>
-<meta charset="UTF-8">
-<title> 회원 정보 수정 및 탈퇴 </title>
+
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <meta name="description" content="">
+    <meta name="author" content="">
+
+    <title>신비한 동물사전</title>
+    
+	<!-- css script -->
+	<%@include file="../fragments/common-css.jsp" %>
+	
+</head>
+
 <script type="text/javascript"> 
 
 function validate() {
@@ -38,16 +50,76 @@ function check(valid, what, message) {
 }
 
 </script>
-<body>
-<form method="post" onsubmit="return validate();">
-	아이디 : <input type="text" name="id" value="${userInfo.id}" readonly /> <br />
-    비밀번호 : <input type="password" name="password" id="password" required/> <br />
-    비밀번호 확인 : <input type="password" name="password2" id="password2" required/> <br />
-    이름 : <input type="text" name="name" value="${userInfo.name}" readonly /> <br />
-    핸드폰번호 : <input type="text" name="phone" value="${userInfo.phone}" readonly /> <br />
-    <input type="submit" value="비밀번호 변경" onclick="modify();" />
-    <button type="button" onclick="location.href='../main'">홈으로</button>
-</form>
+
+<body id="section_1">
+
+	<!-- header -->
+	<%@include file="../fragments/header.jsp" %>
+	
+	<!-- nav -->
+	<%@include file="../fragments/nav.jsp" %>
+	
+	<main>		
+		<section class="text-left" style="margin: auto; padding: 5% 0;">
+			<div class="container">
+			
+				<h3 class="." style="color: var(--primary-color)">내 정보 수정📝</h3>
+				<br>
+				<span>
+					<small class="small-font text-secondary">* 표시는 필수 입력사항 입니다.</small>
+				</span>
+				
+				<br>
+				<br>
+				
+				<form method="post" onsubmit="return validate();">
+
+					<table class="table table-bordered">
+					  <tbody>
+					    <tr>
+					      <td>아이디</td>
+					      <td><input type="text" name="id" value="${userInfo.id}" style="border:0 solid black" readonly /></td>
+						</tr>
+					    <tr>
+					      <td>비밀번호*</td>
+					      <td><input type="password" name="password" id="password" required/></td>
+					    </tr>
+					    <tr>
+					      <td>비밀번호 확인*</td>
+					      <td><input type="password" name="password2" id="password2" required/></td>
+					    </tr>
+					    <tr>
+					      <td>이름</td>
+					      <td><input type="text" name="name" value="${userInfo.name}" style="border:0 solid black" readonly /></td>
+					    </tr>
+					    <tr>
+					      <td>전화번호</td>
+					      <td><input type="text" name="phone" value="${userInfo.phone}" style="border:0 solid black" readonly /></td>
+					    </tr>
+					  </tbody>
+					</table>	
+									
+					<button type="submit" class="btn btn-outline-dark" onclick="modify();">비밀번호 변경</button>
+					<button type="button" class="btn btn-success" onclick="location.href='../main'">메인으로</button>
+					
+					
+				</form>
+
+
+
+			</div>
+		</section>
+	</main>
+
+	<!-- footer -->
+	<%@include file="../fragments/footer.jsp" %>
+
+	<!-- js -->
+	<%@include file="../fragments/common-js.jsp" %>
+	
 </body>
-</head>
+
+
+
+
 </html>
